not delete question from another author

+++ src/domain/forum/application/use-cases/delete-question.ts
@@ -3,6 +3,7 @@ import { QuestionRepository } from "../repositories/questions-repository";
 
 interface DeleteQuestionUseCaseRequest {
   questionId: string;
+  authorId: string;
 }
 
 interface DeleteQuestionUseCaseResponse {}
@@ -10,14 +11,19 @@ interface DeleteQuestionUseCaseResponse {}
 export class DeleteQuestionUseCase {
 
   constructor(private questionRepository: QuestionRepository) {}
-  
+
   async execute({
+    authorId,
     questionId,
   }: DeleteQuestionUseCaseRequest): Promise<DeleteQuestionUseCaseResponse> {
     const question = await this.questionRepository.findById(questionId);
 
     if (!question) {
       throw new Error("Question not found.");
+    }
+
+    if(authorId !== question.authorId.toString()){
+      throw new Error("Not allowed.");
     }
 
     await this.questionRepository.delete(question);
